create instance rule - already exists
created a separate method that, before a new insance is created, checks if the instance already exists. If it does, a new message appears letting the user know. Otherwise, the new instance gets created. In addition, the edit method got an additional rule added to it so that the modal would only open if the section name in question is not empty